add model hooks back to routes
