+++ ui/lib/pki/addon/templates/configuration/edit.hbs
@@ -10,4 +10,4 @@
   </p.levelLeft>
 </PageHeader>
 
-<Page::PkiConfigurationEdit @urls={{this.model.urls}} @crl={{this.model.crl}} @backend={{this.model.engine.id}} />
+<Page::PkiConfigurationEdit @urls={{this.model.urls}} @crl={{this.model.crl}} @backend={{this.model.engineId}} />
