+++ ui/tests/acceptance/pki/pki-engine-route-cleanup-test.js
@@ -299,9 +299,10 @@ module('Acceptance | pki engine route cleanup test', function (hooks) {
       await click(SELECTORS.configuration.generateRootSave);
       // Go to list view so we fetch all the issuers
       await visit(`/vault/secrets/${this.mountPath}/pki/issuers`);
+
       issuers = this.store.peekAll('pki/issuer');
       const issuerId = issuers.objectAt(0).id;
-      assert.strictEqual(issuers.length, 1, 'Issuer exists on model');
+      assert.strictEqual(issuers.length, 1, 'Issuer exists on model in list');
       await visit(`/vault/secrets/${this.mountPath}/pki/issuers/${issuerId}/details`);
       await click(SELECTORS.issuerDetails.configure);
       issuer = this.store.peekRecord('pki/issuer', issuerId);
@@ -310,7 +311,7 @@ module('Acceptance | pki engine route cleanup test', function (hooks) {
       assert.true(issuer.hasDirtyAttributes, 'Model is dirty');
       await click(SELECTORS.overviewBreadcrumb);
       issuers = this.store.peekAll('pki/issuer');
-      assert.strictEqual(issuers.length, 1, 'Issuer exists on model');
+      assert.strictEqual(issuers.length, 1, 'Issuer exists on model in overview');
       issuer = this.store.peekRecord('pki/issuer', issuerId);
       assert.false(issuer.hasDirtyAttributes, 'Dirty attrs were rolled back');
     });
